requirements.txt
streamlit>=1.28.0
requests>=2.28.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pydub>=0.25.1
easyocr>=1.7.0
openai-whisper>=20231117
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
soundfile>=0.12.0
sentence-transformers>=2.2.0
